requirements.txt
md2tgmd
g4f
python-telegram-bot
nest_asyncio